feat: Add a max-sized header and media queries
This one went well! I added a wrapper to the entire
header and used this to center it across the whole
page instead of having the links all the way left or
right. This is what most modern web pages do. Also
the header background does stretch, just not the header
elements.